Move code around according to review
