+++ app/scripts/components/exploration/components/analysis-metrics-dropdown.tsx
