+++ app/scripts/components/exploration/components/timeline/timeline-utils.ts
